+++ R/data_highlights_module.R
@@ -97,8 +97,15 @@ data_highlights_ui <- function(id) {
 data_highlights_server <- function(id) {
   moduleServer(id, function(input, output, session) {
 
-    df <- reactive(session$userData$df$act)
-    df_metadata <- reactive(session$userData$df$act_meta())
+    df <- reactive({
+      req(session$userData$df$act)
+      session$userData$df$act
+    })
+
+    df_metadata <- reactive({
+      req(session$userData$df$act_meta)
+      session$userData$df$act_meta()
+    })
 
     output$var_num_vars <- renderText(sapply(df(), is.numeric) |> sum())
     output$var_char_vars <- renderText(sapply(df(), is.character) |> sum())