+++ tests/testthat/test-spada_server.R
@@ -2,9 +2,25 @@
 
 dfs <- list('df_iris' = iris, 'df_mtcars' = mtcars)
 
+temp_dir <- tempdir()
+
+start_conf <- list(
+  'conf_dir' = paste0(temp_dir, '\\conf'),
+  'data_dir' = paste0(temp_dir, '\\data'),
+  'theme' = 'spada_theme',
+  'file_size' = 1000,
+  'restore_session' = 'never',
+  'save_session' = 'ask',
+  'restore_data_status' = 0,
+  'restore_output_status' = 0,
+  'restore_status' = NULL,
+  'plot_fill_color' = plot_fill_color,
+  'plot_line_color' = plot_line_color
+)
+
 # test inputed datasets -------------------------------------------------------
 test_that('Test inputed datasets', {
-  testServer(spada_server(datasets = dfs), {
+  testServer(spada_server(datasets = dfs, conf = start_conf), {
     expect_equal(session$userData$dt$dt[[1]], iris |> as.data.table())
     expect_equal(session$userData$dt$dt[[2]], mtcars |> as.data.table())
     expect_equal(names(session$userData$dt$dt), c('df_iris', 'df_mtcars'))
@@ -14,7 +30,7 @@ test_that('Test inputed datasets', {
 
 # test check datasets classes -------------------------------------------------
 test_that('Test inputed datasets', {
-  testServer(spada_server(datasets = dfs), {
+  testServer(spada_server(datasets = dfs, conf = start_conf), {
     expect_equal(session$userData$dt$dt[[1]] |> class(), c('data.table', 'data.frame'))
     expect_equal(session$userData$dt$dt[[2]] |> class(), c('data.table', 'data.frame'))
     expect_equal(session$userData$df$act |> class(), c('data.table', 'data.frame'))
@@ -23,7 +39,7 @@ test_that('Test inputed datasets', {
 
 # test active df --------------------------------------------------------------
 test_that('Test active df', {
-  testServer(spada_server(datasets = dfs), {
+  testServer(spada_server(datasets = dfs, conf = start_conf), {
     expect_equal(session$userData$df$act, iris |> as.data.table())
     expect_equal(session$userData$df$act_name, 'df_iris')
     expect_equal(session$userData$df$bkp, NULL)
@@ -32,14 +48,14 @@ test_that('Test active df', {
 
 # test active df metadata -----------------------------------------------------
 test_that('Test active df metadata', {
-  testServer(spada_server(datasets = dfs), {
+  testServer(spada_server(datasets = dfs, conf = start_conf), {
     expect_equal(session$userData$df$act_meta(), df_info(iris))
   })
 })
 
 # test change active df -------------------------------------------------------
 test_that('Test change active df', {
-  testServer(spada_server(datasets = dfs), {
+  testServer(spada_server(datasets = dfs, conf = start_conf), {
 
     session$setInputs(pD_data_sel_df = 'df_mtcars')
 
@@ -53,7 +69,7 @@ test_that('Test change active df', {
 
 # test rename active df -------------------------------------------------------
 test_that('Test rename active df', {
-  testServer(spada_server(datasets = dfs), {
+  testServer(spada_server(datasets = dfs, conf = start_conf), {
     session$setInputs(pD_data_sel_df = 'df_iris')
     session$setInputs(pD_data_txt_new_name = 'new_iris')
 
@@ -64,7 +80,7 @@ test_that('Test rename active df', {
 })
 
 test_that('Test rename active df - invalid name', {
-  testServer(spada_server(datasets = dfs), {
+  testServer(spada_server(datasets = dfs, conf = start_conf), {
     session$setInputs(pD_data_sel_df = 'df_iris')
     session$setInputs(pD_data_txt_new_name = 'new iris')
 
@@ -76,10 +92,10 @@ test_that('Test rename active df - invalid name', {
 
 # test copy df --------------------------------------------------------------
 test_that('Test copy df', {
-  testServer(spada_server(datasets = dfs), {
+  testServer(spada_server(datasets = dfs, conf = start_conf), {
 
     session$setInputs(pD_data_txt_new_name = 'df_iris2')
-
+    session$setInputs(pD_data_sel_df = 'df_iris')
     session$setInputs(pD_data_btn_copy_dataset = 1)
 
     expect_equal(session$userData$dt_names(), c('df_iris', 'df_mtcars', 'df_iris2'))
@@ -90,7 +106,7 @@ test_that('Test copy df', {
 
 # test delete df --------------------------------------------------------------
 test_that('Test delete non active df', {
-  testServer(spada_server(datasets = dfs), {
+  testServer(spada_server(datasets = dfs, conf = start_conf), {
     session$setInputs(pD_data_sel_df = 'df_mtcars')
 
     session$setInputs(pD_data_btn_delete_dataset = 1)
@@ -102,7 +118,7 @@ test_that('Test delete non active df', {
 })
 
 test_that('Test delete active df', {
-  testServer(spada_server(datasets = dfs), {
+  testServer(spada_server(datasets = dfs, conf = start_conf), {
     session$setInputs(pD_data_sel_df = 'df_iris')
 
     session$setInputs(pD_data_btn_delete_dataset = 1)